+++ hydra/detail/functors/ProcessGaussKronrodQuadrature.h
@@ -103,8 +103,7 @@ struct GaussKronrodUnary
 
 		GaussKronrodCall result;
 
-		GReal_t function_call    = abscissa_Weight*(fFunctor(abscissa_X_M)
-				+ fFunctor(abscissa_X_M) ) ;
+		GReal_t function_call    = abscissa_Weight*(fFunctor(abscissa_X_M) + fFunctor(abscissa_X_P) ) ;
 
 		result.fGaussCall        = function_call*rule_Gauss_Weight;
 		result.fGaussKronrodCall = function_call*rule_GaussKronrod_Weight;